Moving d_xp_ to state GPU and doing the copies
I moved d_xp_ Device array to StatePropagatorDataGPU so it can
be moved back and forth to the host during integration instead of relying
on an internal temporary array.

I'm still missing the necessary checks for velocity updates during constraints,
and skipping SETTLE on constrain_coordinates() is the doLincsOnCpu flag is on.