Doing constraints on CPUs for now
For BenchPEP (and probably other molecules), we need to have update
groups enabled on a domain decomposition (multi-rank simulation with
more than 1 PP rank) run.
Now, what we see here is that update groups are not being used, which
breaks the SETTLE code on the GPU.

We need to understand how to get around this and move SETTLE
back to the GPU, by re-enabling SETTLE even when we have all-bonds enabled.
Without update-groups, we probably have some constraint crossing the domain
boundary, which makes SETTLE_GPU break.

Disabling SETTLE_GPU for now if we do not pass the LINCS checks.